requirements.txt
django
gunicorn
whitenoise
dj-database-url
psycopg2-binary
scikit-learn
gtts
xhtml2pdf
pdfplumber
python-docx
python-dotenv
google-genai
